Fixed crash when trying to take a screenshot
Potentially fixed more crashes related to invalid casts.

Major code clean up

+++ Terraria/Terraria/ID/AnimationID.cs
@@ -12,7 +12,7 @@ namespace Terraria.ID
 {
     public class AnimationID
     {
-        public const short MushroomStatueTurningOn = (short)0;
-        public const short MushroomStatueTurningOff = (short)1;
+        public const short MushroomStatueTurningOn = 0;
+        public const short MushroomStatueTurningOff = 1;
     }
 }
