+++ Terraria/Terraria/ID/ChainID.cs
@@ -12,22 +12,22 @@ namespace Terraria.ID
 {
     public static class ChainID
     {
-        public const short TendonHook = (short)0;
-        public const short ThornHook = (short)1;
-        public const short IlluminantHook = (short)2;
-        public const short Wormhook = (short)3;
-        public const short SilkRope = (short)4;
-        public const short SilkRope2 = (short)5;
-        public const short WebRope = (short)6;
-        public const short WebRope2 = (short)7;
-        public const short LunarSolar = (short)8;
-        public const short LunarVortex = (short)9;
-        public const short LunarNebula = (short)10;
-        public const short LunarStardust = (short)11;
-        public const short LunarSolarGlow = (short)12;
-        public const short LunarVortexGlow = (short)13;
-        public const short LunarNebulaGlow = (short)14;
-        public const short LunarStardustGlow = (short)15;
-        public const short Count = (short)16;
+        public const short TendonHook = 0;
+        public const short ThornHook = 1;
+        public const short IlluminantHook = 2;
+        public const short Wormhook = 3;
+        public const short SilkRope = 4;
+        public const short SilkRope2 = 5;
+        public const short WebRope = 6;
+        public const short WebRope2 = 7;
+        public const short LunarSolar = 8;
+        public const short LunarVortex = 9;
+        public const short LunarNebula = 10;
+        public const short LunarStardust = 11;
+        public const short LunarSolarGlow = 12;
+        public const short LunarVortexGlow = 13;
+        public const short LunarNebulaGlow = 14;
+        public const short LunarStardustGlow = 15;
+        public const short Count = 16;
     }
 }
