+++ Terraria/Terraria/ID/InvasionID.cs
@@ -12,12 +12,12 @@ namespace Terraria.ID
 {
     public static class InvasionID
     {
-        public const short CachedInvasions = (short)2;
-        public const short None = (short)0;
-        public const short GoblinArmy = (short)1;
-        public const short SnowLegion = (short)2;
-        public const short PirateInvasion = (short)3;
-        public const short MartianMadness = (short)4;
-        public const short Count = (short)5;
+        public const short CachedInvasions = 2;
+        public const short None = 0;
+        public const short GoblinArmy = 1;
+        public const short SnowLegion = 2;
+        public const short PirateInvasion = 3;
+        public const short MartianMadness = 4;
+        public const short Count = 5;
     }
 }
